separated spec files for header
