+++ cypress/integration/v2/blocks/header/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Header Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
